+++ src/client/src/components/styledTabComponents.tsx
@@ -4,7 +4,7 @@ import { theme } from "../AppTheme";
 
 export const BdmsTabContentBox = styled(Box)(() => ({
   backgroundColor: theme.palette.background.default,
-  border: `1px solid ${theme.palette.boxShadow}`,
+  border: `1px solid ${theme.palette.border.light}`,
   padding: `${theme.spacing(3)}`,
   display: "flex",
   flexDirection: "column",
@@ -35,9 +35,9 @@ export const BdmsTab = styled((props: BdmsTabProps) => <Tab disableRipple {...pr
   "&.Mui-selected": {
     color: theme.palette.background.menuItemActive,
     backgroundColor: theme.palette.background.default,
-    borderTop: `1px solid ${theme.palette.boxShadow}`,
-    borderRight: `1px solid ${theme.palette.boxShadow}`,
-    borderLeft: `1px solid ${theme.palette.boxShadow}`,
+    borderTop: `1px solid ${theme.palette.border.light}`,
+    borderRight: `1px solid ${theme.palette.border.light}`,
+    borderLeft: `1px solid ${theme.palette.border.light}`,
     borderBottom: `1px solid ${theme.palette.background.default}`,
     borderRadius: `${theme.spacing(0.5)} ${theme.spacing(0.5)} 0 0`,
     top: "1px",