+++ src/client/src/pages/detail/form/location/elevationSegment.tsx
@@ -1,9 +1,7 @@
 import { FC } from "react";
 import { UseFormReturn } from "react-hook-form";
-import { TextField } from "@mui/material/";
-import { t } from "i18next";
 import { useDomains } from "../../../../api/fetchApiV2";
-import { FormContainer, FormDomainSelect, FormInput } from "../../../../components/form/form.ts";
+import { FormContainer, FormDomainSelect, FormInput, FormInputDisplayOnly } from "../../../../components/form/form.ts";
 import { Codelist } from "../../../../components/legacyComponents/domain/domainInterface.ts";
 import { FormSegmentBox } from "../../../../components/styledComponents.ts";
 import { LocationBaseProps, LocationFormInputs } from "./locationPanelInterfaces.tsx";
@@ -54,10 +52,8 @@ const ElevationSegment: FC<ElevationSegmentProps> = ({ borehole, formMethods }) 
             schemaName={"reference_elevation_type"}
             selected={borehole.referenceElevationTypeId}
           />
-          <TextField
-            InputProps={{ readOnly: true }}
-            className="readonly"
-            label={t("height_reference_system")}
+          <FormInputDisplayOnly
+            label={"height_reference_system"}
             value={domains?.find((d: Codelist) => d.id === borehole.hrsId)?.en}
           />
         </FormContainer>